+++ src/utils/settings-update.js
@@ -4,7 +4,7 @@
  *
  * @public
  * @name updateSettings
- * @memberof Context.helpers
+ * @memberof Utils
  * @param {Object} ctx The Context Object
  * @param  {Object} newSettings settings to update
  * @return {Promise}